Moved FieldPreview to components root directory for sharing it between components
