+++ src/components/FieldPreview/fieldPreview.m.scss
@@ -1,4 +1,4 @@
-@import 'styles/mixins';
+@import '../../styles/mixins';
 
 .core {
   @include flex-row;
