+++ src/components/HistoryItemDetails/components/index.js
@@ -1,4 +1,3 @@
 // @flow strict
 
 export { AssetItemPreview } from './AssetItemPreview/AssetItemPreview'
-export { FieldPreview } from './FieldPreview/FieldPreview'
