+++ src/components/index.js
@@ -33,6 +33,7 @@ export { TransactionItem } from './TransactionItemNew/TransactionItem'
 export { ButtonWithConfirm } from './ButtonWithConfirm/ButtonWithConfirm'
 export { HistoryItemDetails } from './HistoryItemDetails/HistoryItemDetails'
 export { WalletPasswordForm } from './WalletPasswordForm/WalletPasswordForm'
+export { FieldPreview } from './FieldPreview/FieldPreview'
 
 export {
   QRCode,
